+++ apps/web/pages/apps/index.tsx
@@ -55,9 +55,8 @@ const AppsPageHeader = observer(() => {
   return <PageHeader extra={pageHeaderButtons} ghost={false} title="Apps" />
 })
 
-const AppsPage: CodelabPage<DashboardTemplateProps> = observer(() => {
+const AppsPage: CodelabPage<DashboardTemplateProps> = () => {
   const store = useStore()
-  store.appService.getAll()
 
   return (
     <>
@@ -74,7 +73,7 @@ const AppsPage: CodelabPage<DashboardTemplateProps> = observer(() => {
       </ContentSection>
     </>
   )
-})
+}
 
 export default AppsPage
 
@@ -84,7 +83,6 @@ export const getServerSideProps = withPageAuthRequired<Snapshot>({
     await setClientAuthHeaders(context)
 
     const store = initializeStore()
-
     await store.appService.getAll()
 
     return {
