+++ libs/frontend/modules/app/src/use-cases/create-app/createAppSchema.ts
@@ -14,4 +14,4 @@ export const createAppSchema: JSONSchemaType<CreateAppInput> = {
     },
   },
   required: ['name'],
-}
+} as const
