+++ libs/frontend/view/components/form/index.ts
@@ -1,11 +1,11 @@
 export * from './components/Form'
 export * from './components/FormModal'
-export * from './components/ModalForm'
 export * from './components/utils'
 export * from './fields/AutoCompleteField'
 export * from './fields/DisplayIfField'
 export * from './fields/UploadField'
 export * from './hooks/submitController'
 export * from './hooks/uniformUtils'
+export * from './modal'
 export * from './monaco/monacoFieldFactory'
 export * from './schema/emptyJsonSchema'
