+++ libs/frontend/view/components/form/modal/index.tsx
@@ -1,0 +1,4 @@
+import { Form } from './ModalForm.Form'
+import { Modal } from './ModalForm.Modal'
+
+export const ModalForm = { Form, Modal }
